// src/app/features/product/models/productList.ts
// Generated by https://quicktype.io

export interface ProductItem {
    id:              number;
    supplierId:      number;
    categoryId:      number;
    quantityPerUnit: string;
    unitPrice:       number;
    unitsInStock:    number;
    unitsOnOrder:    number;
    reorderLevel:    number;
    discontinued:    boolean;
    name:            string;
}