// src/app/features/product/models/productDetails.ts
// Generated by https://quicktype.io

export interface ProductDetails {
    id:              number;
    supplierId:      number;
    categoryId:      number;
    quantityPerUnit: string;
    unitPrice:       number;
    unitsInStock:    number;
    unitsOnOrder:    number;
    reorderLevel:    number;
    discontinued:    boolean;
    name:            string;
}